+++ freqtrade/rpc/api_server/api_schemas.py
