+++ freqtrade/rpc/api_server/api_v1.py
@@ -10,14 +10,14 @@ from freqtrade.constants import USERPATH_STRATEGIES
 from freqtrade.data.history import get_datahandler
 from freqtrade.exceptions import OperationalException
 from freqtrade.rpc import RPC
-from freqtrade.rpc.api_server.api_models import (AvailablePairs, Balances, BlacklistPayload,
-                                                 BlacklistResponse, Count, Daily, DeleteTrade,
-                                                 ForceBuyPayload, ForceSellPayload, Locks, Logs,
-                                                 OpenTradeSchema, PairHistory, PerformanceEntry,
-                                                 Ping, PlotConfig, Profit, ResultMsg, Stats,
-                                                 StatusMsg, StrategyListResponse, StrategyResponse,
-                                                 TradeResponse, TradeSchema, Version,
-                                                 WhitelistResponse)
+from freqtrade.rpc.api_server.api_schemas import (AvailablePairs, Balances, BlacklistPayload,
+                                                  BlacklistResponse, Count, Daily, DeleteTrade,
+                                                  ForceBuyPayload, ForceSellPayload, Locks, Logs,
+                                                  OpenTradeSchema, PairHistory, PerformanceEntry,
+                                                  Ping, PlotConfig, Profit, ResultMsg, Stats,
+                                                  StatusMsg, StrategyListResponse, StrategyResponse,
+                                                  TradeResponse, TradeSchema, Version,
+                                                  WhitelistResponse)
 from freqtrade.rpc.api_server.deps import get_config, get_rpc
 from freqtrade.rpc.rpc import RPCException
